+++ source/vulkan/ui.d
@@ -95,7 +95,7 @@ abstract class UiWidget
         auto childContext = localContext.offset(childOffsetX, childOffsetY);
         foreach (index, child; children)
         {
-            childContext.depthBase = localContext.depthBase - cast(float)index * 0.02f;
+            childContext.depthBase = localContext.depthBase - cast(float)index * 0.001f;
             child.render(childContext);
         }
     }